+++ apps/router-demo/router-host-2000/src/runtime-plugin/shared-strategy.ts
@@ -20,4 +20,5 @@ const sharedStrategy: () => FederationRuntimePlugin = () => ({
     return args;
   },
 });
+
 export default sharedStrategy;