+++ apps/router-demo/router-remote1-2001/rsbuild.config.ts
@@ -10,16 +10,15 @@ export default defineConfig({
     alias: {
       react: path.resolve(__dirname, 'node_modules/react'),
       'react-dom': path.resolve(__dirname, 'node_modules/react-dom'),
-      // 'react-router-dom': path.resolve(__dirname, 'node_modules/react-router-dom'),
-      // 'react-router-dom$': path.resolve(
-      //   __dirname,
-      //   'node_modules/@module-federation/bridge-react/dist/router.es.js',
-      // ),
+      // set `react-router-dom/` to reference react-router-dom v5 which shoule be find in node_modules/react-router-dom, otherwise it will cause app.tsx fail to work which in react-router-dom v5 mode.
+      'react-router-dom': path.resolve(
+        __dirname,
+        'node_modules/react-router-dom',
+      ),
     },
   },
   server: {
     port: 2001,
-    host: 'localhost',
   },
   dev: {
     // It is necessary to configure assetPrefix, and in the production environment, you need to configure output.assetPrefix