+++ apps/modernjs-ssr/nested-remote/CHANGELOG.md
@@ -1,5 +1,17 @@
 # modernjs-ssr-nested-remote
 
+## 0.1.18
+
+### Patch Changes
+
+- @module-federation/modern-js@0.5.1
+
+## 0.1.17
+
+### Patch Changes
+
+- @module-federation/modern-js@0.5.0
+
 ## 0.1.16
 
 ### Patch Changes