+++ packages/bridge/bridge-shared/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @module-federation/bridge-shared
 
+## 0.5.1
+
+## 0.5.0
+
+### Patch Changes
+
+- 49d6135: feat(@module-federation/bridge): enhance Bridge capabilities and fix some issues
+
 ## 0.4.0
 
 ## 0.3.5