+++ packages/dts-plugin/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @module-federation/dts-plugin
 
+## 0.5.1
+
+### Patch Changes
+
+- @module-federation/sdk@0.5.1
+- @module-federation/managers@0.5.1
+- @module-federation/third-party-dts-extractor@0.5.1
+
+## 0.5.0
+
+### Patch Changes
+
+- Updated dependencies [8378a77]
+  - @module-federation/sdk@0.5.0
+  - @module-federation/managers@0.5.0
+  - @module-federation/third-party-dts-extractor@0.5.0
+
 ## 0.4.0
 
 ### Patch Changes