+++ packages/esbuild/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @module-federation/esbuild
 
+## 0.0.16
+
+### Patch Changes
+
+- @module-federation/sdk@0.5.1
+
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [8378a77]
+  - @module-federation/sdk@0.5.0
+
 ## 0.0.14
 
 ### Patch Changes