+++ packages/nextjs-mf/CHANGELOG.md
@@ -1,5 +1,27 @@
 # [8.1.0-canary.7](https://github.com/module-federation/core/compare/nextjs-mf-8.1.0-canary.6...nextjs-mf-8.1.0-canary.7) (2023-11-21)
 
+## 8.4.11
+
+### Patch Changes
+
+- @module-federation/enhanced@0.5.1
+- @module-federation/node@2.5.8
+- @module-federation/runtime@0.5.1
+- @module-federation/sdk@0.5.1
+- @module-federation/utilities@3.1.4
+
+## 8.4.10
+
+### Patch Changes
+
+- Updated dependencies [8378a77]
+- Updated dependencies [5c7ac8a]
+  - @module-federation/runtime@0.5.0
+  - @module-federation/sdk@0.5.0
+  - @module-federation/enhanced@0.5.0
+  - @module-federation/node@2.5.7
+  - @module-federation/utilities@3.1.3
+
 ## 8.4.9
 
 ### Patch Changes