+++ packages/third-party-dts-extractor/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @module-federation/third-party-dts-extractor
 
+## 0.5.1
+
+## 0.5.0
+
 ## 0.4.0
 
 ## 0.3.5